refactor(factory-definitions): use Models backend registry predicate
