+++ pkg/services/factory_definitions/internal/services/validation/impl/topology_validator_test.go
@@ -586,71 +586,6 @@ func TestRuleResourceDefinitions_AcceptsModelResourceMetadata(t *testing.T) {
 	}
 }
 
-func TestManagedRuntimeDependencyTargets_CharacterizesCurrentMembership(t *testing.T) {
-	t.Parallel()
-
-	tests := []struct {
-		name        string
-		model       string
-		backend     string
-		wantCode    string
-		wantMessage string
-	}{
-		{
-			name:    "normalized OmniVoice and LlamaCpp pairing",
-			model:   "  omnivoice_q4_k_m ",
-			backend: "  llamaCpp \t",
-		},
-		{
-			name:        "known identity with different backend",
-			model:       "OMNIVOICE_Q4_K_M",
-			backend:     "GGUF",
-			wantCode:    CodeManagedRuntimeInvalidBackend,
-			wantMessage: `requires backend "LLAMACPP"`,
-		},
-		{
-			name:        "unknown identity with managed backend",
-			model:       "UNKNOWN_MODEL",
-			backend:     "LLAMACPP",
-			wantCode:    CodeManagedRuntimeUnsupportedIdentity,
-			wantMessage: "is not supported in this environment",
-		},
-	}
-	for _, testCase := range tests {
-		testCase := testCase
-		t.Run(testCase.name, func(t *testing.T) {
-			t.Parallel()
-			cfg := &factorydefinitions.FactoryConfig{
-				Resources: []factorydefinitions.ResourceConfig{{
-					Name:       "voice-assets",
-					Type:       factorydefinitions.ResourceTypeModel,
-					Capacity:   1,
-					Model:      testCase.model,
-					Backend:    testCase.backend,
-					LoadPolicy: "ON_DEMAND",
-				}},
-			}
-
-			targets := ManagedRuntimeDependencyTargets(cfg)
-			if testCase.wantCode == "" {
-				if len(targets) != 0 {
-					t.Fatalf("targets = %#v, want no findings", targets)
-				}
-				return
-			}
-			if len(targets) != 1 {
-				t.Fatalf("targets = %#v, want one finding", targets)
-			}
-			if targets[0].Code != testCase.wantCode {
-				t.Fatalf("target code = %q, want %q", targets[0].Code, testCase.wantCode)
-			}
-			if !strings.Contains(targets[0].Message, testCase.wantMessage) {
-				t.Fatalf("target message = %q, want substring %q", targets[0].Message, testCase.wantMessage)
-			}
-		})
-	}
-}
-
 func TestRuleRequiredTools_MissingNameAndCommand(t *testing.T) {
 	cfg := testBaseConfig()
 	cfg.ResourceManifest = &factorydefinitions.PortableResourceManifestConfig{
