+++ pkg/services/models/root_authority_seal_characterization_test.go
@@ -61,34 +61,6 @@ func TestRootServiceAuthority_FakePeerGetModelNotFound(t *testing.T) {
 	}
 }
 
-func TestManagedRuntimeBackendPredicate_UsesOnlyEnabledAliases(t *testing.T) {
-	t.Parallel()
-
-	tests := []struct {
-		name    string
-		backend string
-		want    bool
-	}{
-		{name: "canonical", backend: "LLAMACPP", want: true},
-		{name: "case and surrounding whitespace", backend: "  llamaCpp \t", want: true},
-		{name: "blank", backend: "", want: false},
-		{name: "only whitespace", backend: " \t\n", want: false},
-		{name: "unknown", backend: "GGUF", want: false},
-		{name: "llamacpp artifact identifier", backend: "localai-llamacpp", want: false},
-		{name: "whisper artifact identifier", backend: "localai-whisper", want: false},
-		{name: "vibevoice artifact identifier", backend: "localai-vibevoice", want: false},
-	}
-	for _, testCase := range tests {
-		testCase := testCase
-		t.Run(testCase.name, func(t *testing.T) {
-			t.Parallel()
-			if got := models.IsManagedRuntimeBackend(testCase.backend); got != testCase.want {
-				t.Fatalf("IsManagedRuntimeBackend(%q) = %t, want %t", testCase.backend, got, testCase.want)
-			}
-		})
-	}
-}
-
 func TestRootServiceAuthority_AggregateSurfaceRemainsOnSingularService(t *testing.T) {
 	t.Parallel()
